Refactor main logic into dedicated modules
