+++ src/main.js
@@ -1,861 +1,5 @@
-import { FLEET, USERS, setFleet, setLocations, setUsers, resetToDefaults } from './data.js';
-import { state } from './state.js';
-import { $, $$, fmtDate, showToast, openModal, closeModals, kv, formatOdoLabel } from './utils.js';
-import { populateLocationFilters, renderFleet } from './modules/fleet.js';
-import { renderActivity } from './modules/activity.js';
-import { renderAccountRequests, renderUsers } from './modules/users.js';
-import { openDetail, setDetailTab } from './modules/detail.js';
-import { canViewFleetAsset } from './modules/access.js';
-import { setMainTab } from './modules/navigation.js';
-import { resolveEnvironment } from './environment.js';
-
-function getFleetSummaryById(fleetId) {
-  if (!fleetId) return null;
-  const matches = FLEET.filter(item => item?.fleetId === fleetId);
-  if (!matches.length) {
-    return null;
-  }
-
-  const locations = Array.from(new Set(matches.map(item => item.location).filter(Boolean)));
-  return {
-    id: fleetId,
-    name: matches[0]?.fleetName || '—',
-    locations
-  };
-}
-
-function switchMainTab(tab) {
-  const allowedTabs = Array.isArray(state.allowedTabs) ? state.allowedTabs : ['vloot', 'activiteit', 'users'];
-
-  if (!allowedTabs.length) {
-    state.activeTab = null;
-    setMainTab(null);
-    return;
-  }
-
-  const fallback = allowedTabs[0] || 'vloot';
-  const targetTab = allowedTabs.includes(tab) ? tab : fallback;
-  state.activeTab = targetTab;
-  setMainTab(targetTab);
-}
-
-function applyEnvironmentForRole(role) {
-  const environment = resolveEnvironment(role);
-  state.activeEnvironmentKey = environment.key;
-  state.allowedTabs = [...environment.allowedTabs];
-
-  const nameEl = $('#environmentName');
-  if (nameEl) {
-    nameEl.textContent = environment.label;
-  }
-
-  const badgeEl = $('#environmentBadge');
-  if (badgeEl) {
-    badgeEl.textContent = environment.label;
-  }
-
-  const summaryEl = $('#environmentSummary');
-  if (summaryEl) {
-    summaryEl.textContent = environment.summary;
-  }
-
-  $$('#mainTabs button').forEach(button => {
-    const allowed = environment.allowedTabs.includes(button.dataset.tab);
-    const listItem = button.closest('li');
-    if (listItem) {
-      listItem.classList.toggle('hidden', !allowed);
-    }
-    button.disabled = !allowed;
-    button.classList.toggle('opacity-40', !allowed);
-  });
-
-  switchMainTab(state.activeTab);
-  renderAccountRequests();
-}
-import {
-  fetchFleet,
-  fetchLocations,
-  fetchUsers,
-  fetchProfileByAuthUserId,
-  fetchFleetMemberships,
-  fetchAccountRequests,
-  createAccountRequest,
-  approveAccountRequest,
-  rejectAccountRequest,
-  getCurrentSession,
-  onAuthStateChange,
-  signInWithPassword,
-  signOut as supabaseSignOut,
-  touchProfileSignIn
-} from './api/supabase.js';
-
-function wireEvents() {
-  if (state.eventsWired) return;
-  state.eventsWired = true;
-
-  function closeKebabMenus(except = null) {
-    $$('.kebab-menu').forEach(menu => {
-      if (menu !== except) {
-        menu.classList.add('hidden');
-      }
-    });
-  }
-
-  $('#userBtn').addEventListener('click', () => $('#userMenu').classList.toggle('hidden'));
-  $('#cancelUserMenu').addEventListener('click', () => $('#userMenu').classList.add('hidden'));
-  document.addEventListener('click', event => {
-    if (!event.target.closest('#userBtn') && !event.target.closest('#userMenu')) {
-      $('#userMenu').classList.add('hidden');
-    }
-  });
-  $('#logoutBtn').addEventListener('click', async () => {
-    try {
-      await supabaseSignOut();
-      showToast('Uitgelogd');
-    } catch (error) {
-      console.error('Uitloggen mislukt', error);
-      showToast('Uitloggen mislukt. Probeer opnieuw.');
-    }
-  });
-
-  const loginForm = $('#loginForm');
-  if (loginForm) {
-    loginForm.addEventListener('submit', handleLoginSubmit);
-  }
-
-  $$('#mainTabs button').forEach(button =>
-    button.addEventListener('click', () => {
-      if (button.disabled) return;
-      switchMainTab(button.dataset.tab);
-    })
-  );
-
-  $('#openAccountRequest')?.addEventListener('click', () => {
-    const form = $('#accountRequestForm');
-    if (form?.reset) {
-      form.reset();
-    }
-    openModal('#modalAccountRequest');
-  });
-
-  const accountRequestForm = $('#accountRequestForm');
-  if (accountRequestForm) {
-    accountRequestForm.addEventListener('submit', handleAccountRequestSubmit);
-  }
-
-  $('#locationFilter').addEventListener('change', event => {
-    state.fleetFilter.location = event.target.value;
-    renderFleet();
-  });
-  const applyFleetSearch = () => {
-    state.fleetFilter.query = $('#searchInput').value;
-    renderFleet();
-  };
-
-  $('#searchBtn').addEventListener('click', applyFleetSearch);
-  $('#searchInput').addEventListener('keydown', event => {
-    if (event.key === 'Enter') {
-      event.preventDefault();
-      applyFleetSearch();
-    }
-  });
-
-  document.addEventListener('click', event => {
-    const trigger = event.target.closest('[data-open-detail]');
-    if (trigger) {
-      openDetail(trigger.dataset.openDetail);
-    }
-  });
-
-  document.addEventListener('click', event => {
-    const button = event.target.closest('.kebab > button');
-    if (button) {
-      const menu = button.parentElement.querySelector('.kebab-menu');
-      const shouldShow = menu && menu.classList.contains('hidden');
-      closeKebabMenus(menu);
-      if (menu && shouldShow) {
-        menu.classList.remove('hidden');
-      }
-      return;
-    }
-
-    if (!event.target.closest('.kebab-menu')) {
-      closeKebabMenus();
-    }
-  });
-
-  document.addEventListener('click', event => {
-    const actionButton = event.target.closest('[data-action]');
-    if (!actionButton) return;
-
-    closeKebabMenus();
-
-    const action = actionButton.dataset.action;
-    const id = actionButton.dataset.id || state.selectedTruckId;
-    const truck = FLEET.find(item => item.id === id);
-    if (!truck || !canViewFleetAsset(truck)) return;
-
-    if (action === 'newTicket' || action === 'newTicketFromDetail') {
-      $('#ticketTruck').value = truck.id;
-      openModal('#modalTicket');
-      return;
-    }
-
-    if (action === 'updateOdo' || action === 'updateOdoFromDetail') {
-      $('#odoCurrent').textContent = formatOdoLabel(truck.odo, truck.odoDate);
-      $('#odoNew').value = '';
-      $('#odoSave').dataset.id = id;
-      openModal('#modalOdo');
-      return;
-    }
-
-    if (action === 'editRef' || action === 'editRefFromDetail') {
-      $('#refCurrent').textContent = truck.ref;
-      $('#refNew').value = truck.ref;
-      $('#refSave').dataset.id = id;
-      openModal('#modalRef');
-      return;
-    }
-
-    if (action === 'showContract') {
-      const contract = truck.contract;
-      $('#contractBody').innerHTML = `
-        ${kv('Contractnummer', contract.nummer)}
-        ${kv('Start datum', fmtDate(contract.start))}
-        ${kv('Eind datum', fmtDate(contract.eind))}
-        ${kv('Uren per jaar', contract.uren)}
-        ${kv('Contract type', contract.type)}
-        ${kv('Model', contract.model)}
-      `;
-      openModal('#modalContract');
-      return;
-    }
-
-    if (action === 'inactive') {
-      $('#inactiveDate').valueAsDate = new Date();
-      $('#inactiveConfirm').dataset.id = id;
-      openModal('#modalInactive');
-    }
-  });
-
-  document.addEventListener('click', event => {
-    const requestButton = event.target.closest('[data-request-action]');
-    if (!requestButton) return;
-    handleAccountRequestAction(requestButton);
-  });
-
-  document.querySelectorAll('.modal [data-close]').forEach(button => button.addEventListener('click', closeModals));
-  document.addEventListener('keydown', event => {
-    if (event.key === 'Escape') closeModals();
-  });
-
-  $('#btnNewTicket').addEventListener('click', () => openModal('#modalTicket'));
-  $('#ticketCreate').addEventListener('click', () => {
-    const id = $('#ticketTruck').value;
-    const type = $('#ticketType').value;
-    const desc = $('#ticketDesc').value.trim();
-    if (!id || !desc) {
-      showToast('Vul minimaal truck en omschrijving in.');
-      return;
-    }
-
-    const truck = FLEET.find(item => item.id === id);
-    if (!truck || !canViewFleetAsset(truck)) {
-      showToast('U heeft geen toegang tot dit object.');
-      return;
-    }
-    truck.activity.push({
-      id: `M-${Math.floor(1000 + Math.random() * 9000)}`,
-      type,
-      desc,
-      status: 'Open',
-      date: new Date().toISOString()
-    });
-
-    closeModals();
-    renderFleet();
-    renderActivity();
-    switchMainTab('activiteit');
-    showToast('Servicemelding aangemaakt.');
-
-    $('#ticketOrder').value = '';
-    $('#ticketDesc').value = '';
-    $('#ticketPhotos').value = '';
-  });
-
-  $('#odoSave').addEventListener('click', event => {
-    const id = event.target.dataset.id;
-    const truck = FLEET.find(item => item.id === id);
-    if (!truck || !canViewFleetAsset(truck)) {
-      showToast('U heeft geen toegang tot dit object.');
-      return;
-    }
-    const value = parseInt($('#odoNew').value, 10);
-    const currentOdo = typeof truck.odo === 'number' && Number.isFinite(truck.odo) ? truck.odo : null;
-    if (Number.isNaN(value) || value < 0 || (currentOdo !== null && value < currentOdo)) {
-      showToast('Nieuwe tellerstand moet ≥ huidige zijn.');
-      return;
-    }
-    truck.odo = value;
-    truck.odoDate = new Date().toISOString();
-    closeModals();
-    renderFleet();
-    setDetailTab('info');
-    showToast('Tellerstand bijgewerkt.');
-  });
-
-  $('#refSave').addEventListener('click', event => {
-    const id = event.target.dataset.id;
-    const truck = FLEET.find(item => item.id === id);
-    if (!truck || !canViewFleetAsset(truck)) {
-      showToast('U heeft geen toegang tot dit object.');
-      return;
-    }
-    const value = $('#refNew').value.trim();
-    if (!value) {
-      showToast('Referentie mag niet leeg zijn.');
-      return;
-    }
-    truck.ref = value;
-    closeModals();
-    renderFleet();
-    setDetailTab('info');
-    showToast('Referentie bijgewerkt.');
-  });
-
-  $('#inactiveConfirm').addEventListener('click', event => {
-    const id = event.target.dataset.id;
-    const date = $('#inactiveDate').value;
-    const reason = $('#inactiveReason').value.trim();
-    if (!date || !reason) {
-      showToast('Datum en reden zijn verplicht.');
-      return;
-    }
-    const truck = FLEET.find(item => item.id === id);
-    if (!truck || !canViewFleetAsset(truck)) {
-      showToast('U heeft geen toegang tot dit object.');
-      return;
-    }
-    truck.active = false;
-    closeModals();
-    renderFleet();
-    renderActivity();
-    showToast('Truck gemarkeerd als inactief.');
-  });
-
-  $('#activityLocationFilter').addEventListener('change', renderActivity);
-  const applyActivitySearch = () => {
-    renderActivity();
-  };
-
-  $('#activitySearchBtn').addEventListener('click', applyActivitySearch);
-  $('#activitySearchInput').addEventListener('keydown', event => {
-    if (event.key === 'Enter') {
-      event.preventDefault();
-      applyActivitySearch();
-    }
-  });
-
-  $('#btnAddUser').addEventListener('click', () => {
-    state.editUserId = null;
-    $('#userModalTitle').textContent = 'Gebruiker toevoegen';
-    $('#userName').value = '';
-    $('#userEmail').value = '';
-    $('#userPhone').value = '';
-    $('#userRole').value = 'Gebruiker';
-    openModal('#modalUser');
-  });
-
-  document.addEventListener('click', event => {
-    const button = event.target.closest('[data-user-action]');
-    if (!button) return;
-
-    closeKebabMenus();
-
-    const id = button.dataset.id;
-    if (button.dataset.userAction === 'edit') {
-      state.editUserId = id;
-      const user = USERS.find(item => item.id === id);
-      $('#userModalTitle').textContent = 'Gebruiker bewerken';
-      $('#userName').value = user.name;
-      $('#userEmail').value = user.email;
-      $('#userPhone').value = user.phone;
-      $('#userLocation').value = user.location;
-      $('#userRole').value = user.role;
-      openModal('#modalUser');
-    } else if (button.dataset.userAction === 'delete') {
-      const index = USERS.findIndex(item => item.id === id);
-      if (index > -1) {
-        USERS.splice(index, 1);
-        renderUsers();
-        showToast('Gebruiker verwijderd.');
-      }
-    }
-  });
-
-  $('#userSave').addEventListener('click', () => {
-    const name = $('#userName').value.trim();
-    const email = $('#userEmail').value.trim();
-    const phone = $('#userPhone').value.trim();
-    const location = $('#userLocation').value;
-    const role = $('#userRole').value;
-
-    if (!name || !email) {
-      showToast('Naam en e-mail zijn verplicht.');
-      return;
-    }
-
-    if (state.editUserId) {
-      const user = USERS.find(item => item.id === state.editUserId);
-      Object.assign(user, { name, email, phone, location, role });
-      showToast('Gebruiker bijgewerkt.');
-    } else {
-      USERS.push({ id: `U${Math.floor(Math.random() * 10000)}`, name, email, phone, location, role });
-      showToast('Gebruiker toegevoegd.');
-    }
-
-    closeModals();
-    renderUsers();
-  });
-
-  $('#usersPrev').addEventListener('click', () => {
-    if (state.usersPage > 1) {
-      state.usersPage -= 1;
-      renderUsers();
-    }
-  });
-
-  $('#usersNext').addEventListener('click', () => {
-    state.usersPage += 1;
-    renderUsers();
-  });
-
-  $('#usersPageSize').addEventListener('change', () => {
-    state.usersPage = 1;
-    renderUsers();
-  });
-
-  $('#backToFleet').addEventListener('click', () => {
-    switchMainTab('vloot');
-  });
-
-  $$('#truckDetail [data-subtab]').forEach(button => button.addEventListener('click', () => setDetailTab(button.dataset.subtab)));
-}
-
-function setLoginStatus(message = '') {
-  const statusEl = $('#loginStatus');
-  if (!statusEl) return;
-  statusEl.textContent = message;
-  statusEl.classList.toggle('hidden', !message);
-}
-
-function setLoginError(message = '') {
-  const errorEl = $('#loginError');
-  if (!errorEl) return;
-  errorEl.textContent = message;
-  errorEl.classList.toggle('hidden', !message);
-}
-
-function setLoginFormDisabled(disabled) {
-  const controls = ['#loginEmail', '#loginPassword', '#loginSubmit'].map(selector => $(selector));
-  controls.forEach(control => {
-    if (control) {
-      control.disabled = disabled;
-    }
-  });
-
-  const defaultLabel = $('#loginSubmitDefault');
-  const loadingLabel = $('#loginSubmitLoading');
-  if (defaultLabel && loadingLabel) {
-    defaultLabel.classList.toggle('hidden', disabled);
-    loadingLabel.classList.toggle('hidden', !disabled);
-  }
-}
-
-function showLoginPage() {
-  $('#loginPage')?.classList.remove('hidden');
-  $('#app')?.classList.add('hidden');
-  setLoginStatus('');
-  setLoginError('');
-  setLoginFormDisabled(false);
-  const emailInput = $('#loginEmail');
-  if (emailInput) {
-    emailInput.focus();
-  }
-  const passwordInput = $('#loginPassword');
-  if (passwordInput) {
-    passwordInput.value = '';
-  }
-}
-
-function showAppShell() {
-  $('#app')?.classList.remove('hidden');
-  $('#loginPage')?.classList.add('hidden');
-}
-
-async function handleLoginSubmit(event) {
-  event.preventDefault();
-
-  const email = $('#loginEmail')?.value.trim();
-  const password = $('#loginPassword')?.value;
-
-  if (!email || !password) {
-    setLoginError('Vul zowel e-mailadres als wachtwoord in.');
-    return;
-  }
-
-  setLoginError('');
-  setLoginStatus('Bezig met inloggen…');
-  setLoginFormDisabled(true);
-
-  try {
-    await signInWithPassword({ email, password });
-  } catch (error) {
-    console.error('Inloggen mislukt', error);
-    const message =
-      error?.message === 'Invalid login credentials'
-        ? 'Onjuiste inloggegevens. Controleer e-mailadres en wachtwoord.'
-        : 'Inloggen mislukt. Probeer het opnieuw.';
-    setLoginError(message);
-    setLoginStatus('');
-    setLoginFormDisabled(false);
-  }
-}
-
-async function handleAccountRequestSubmit(event) {
-  event.preventDefault();
-
-  const name = $('#requestName')?.value.trim();
-  const organisation = $('#requestOrganisation')?.value.trim();
-  const email = $('#requestEmail')?.value.trim();
-  const phone = $('#requestPhone')?.value.trim();
-  const requestNotes = $('#requestNotes')?.value.trim();
-
-  if (!name || !organisation || !email) {
-    showToast('Naam, organisatie en e-mailadres zijn verplicht.');
-    return;
-  }
-
-  const form = event.target;
-  const submitButton = form.querySelector('[type="submit"]');
-  if (submitButton) {
-    submitButton.disabled = true;
-  }
-
-  try {
-    const createdRequest = await createAccountRequest({
-      name,
-      organisation,
-      email,
-      phone,
-      requestNotes
-    });
-
-    if (typeof form.reset === 'function') {
-      form.reset();
-    }
-
-    closeModals();
-
-    if (Array.isArray(state.accountRequests) && state.allowedTabs?.includes('users') && createdRequest) {
-      state.accountRequests.unshift(createdRequest);
-      renderAccountRequests();
-    }
-
-    showToast(
-      'Uw aanvraag is ontvangen. U kunt direct een wachtwoord instellen en inloggen; alle inhoud blijft verborgen totdat een beheerder uw rol heeft toegewezen en het account heeft goedgekeurd.'
-    );
-  } catch (error) {
-    console.error('Accountaanvraag versturen mislukt', error);
-    showToast('Het versturen van de accountaanvraag is mislukt. Probeer het opnieuw.');
-  } finally {
-    if (submitButton) {
-      submitButton.disabled = false;
-    }
-  }
-}
-
-async function handleAccountRequestAction(button) {
-  const action = button.dataset.requestAction;
-  const requestId = button.dataset.id;
-  if (!action || !requestId) return;
-
-  const request = state.accountRequests.find(item => item.id === requestId);
-  if (!request) return;
-
-  const container = button.closest('[data-request]');
-  if (!container) return;
-
-  const toggleButtonsDisabled = disabled => {
-    container.querySelectorAll('[data-request-action]').forEach(actionButton => {
-      actionButton.disabled = disabled;
-    });
-  };
-
-  if (action === 'reject') {
-    toggleButtonsDisabled(true);
-    try {
-      const updatedRequest = await rejectAccountRequest({
-        id: requestId,
-        assignedByProfileId: state.profile?.id || null
-      });
-
-      if (updatedRequest) {
-        Object.assign(request, updatedRequest);
-        renderAccountRequests();
-        showToast(`Aanvraag van ${request.name} geweigerd.`);
-      }
-    } catch (error) {
-      console.error('Accountaanvraag weigeren mislukt', error);
-      showToast('Aanvraag kon niet worden geweigerd. Probeer het opnieuw.');
-    } finally {
-      toggleButtonsDisabled(false);
-    }
-    return;
-  }
-
-  if (action !== 'approve') {
-    return;
-  }
-
-  const roleSelect = container.querySelector('[data-request-role]');
-  const fleetSelect = container.querySelector('[data-request-fleet]');
-  const selectedRole = roleSelect?.value || 'Gebruiker';
-  const selectedFleetId = fleetSelect?.value || '';
-  const fleetSummary = getFleetSummaryById(selectedFleetId);
-  const newUserLocation = fleetSummary?.locations?.[0] || request.organisation || '—';
-
-  toggleButtonsDisabled(true);
-
-  try {
-    const updatedRequest = await approveAccountRequest({
-      id: requestId,
-      assignedRole: selectedRole,
-      assignedFleetId: selectedFleetId || null,
-      assignedByProfileId: state.profile?.id || null
-    });
-
-    if (updatedRequest) {
-      Object.assign(request, updatedRequest);
-
-      USERS.push({
-        id: `U${Math.floor(Math.random() * 10000)}`,
-        name: request.name,
-        email: request.email,
-        phone: request.phone || '',
-        location: newUserLocation,
-        role: selectedRole
-      });
-
-      renderUsers();
-      renderAccountRequests();
-      showToast(`Account toegekend aan ${request.name}.`);
-    }
-  } catch (error) {
-    console.error('Accountaanvraag toekennen mislukt', error);
-    showToast('Account kon niet worden toegekend. Probeer het opnieuw.');
-  } finally {
-    toggleButtonsDisabled(false);
-  }
-}
-
-async function handleAuthenticatedSession(session, { forceReload = false } = {}) {
-  const currentToken = state.session?.access_token ?? null;
-  const nextToken = session?.access_token ?? null;
-
-  if (!forceReload && currentToken === nextToken) {
-    return;
-  }
-
-  state.session = session;
-
-  if (!session) {
-    state.profile = null;
-    state.hasLoadedInitialData = false;
-    state.accessibleFleetIds = null;
-    resetToDefaults();
-    state.fleetFilter.location = 'Alle locaties';
-    state.fleetFilter.query = '';
-    state.selectedTruckId = null;
-    state.usersPage = 1;
-    state.usersPageSize = 10;
-    state.editUserId = null;
-    state.allowedTabs = [];
-    state.activeEnvironmentKey = 'pending';
-    state.activeTab = null;
-    setMainTab(null);
-    $('#currentUserName').textContent = 'Niet ingelogd';
-    const userMenuName = $('#userMenuName');
-    if (userMenuName) {
-      userMenuName.textContent = 'Niet ingelogd';
-    }
-    const userMenuEmail = $('#userMenuEmail');
-    if (userMenuEmail) {
-      userMenuEmail.textContent = '';
-    }
-    showLoginPage();
-    return;
-  }
-
-  setLoginStatus('Gegevens worden geladen…');
-
-  let profile = state.profile;
-  const shouldFetchProfile = forceReload || !profile || profile?.auth_user_id !== session.user.id;
-
-  if (shouldFetchProfile) {
-    try {
-      profile = await fetchProfileByAuthUserId(session.user.id);
-    } catch (error) {
-      console.error('Kon profielgegevens niet laden', error);
-    }
-  }
-
-  state.profile = profile ?? null;
-
-  const displayName =
-    profile?.display_name || session.user.user_metadata?.full_name || session.user.email || 'Ingelogde gebruiker';
-  $('#currentUserName').textContent = displayName;
-  const userMenuName = $('#userMenuName');
-  if (userMenuName) {
-    userMenuName.textContent = displayName;
-  }
-  const userMenuEmail = $('#userMenuEmail');
-  if (userMenuEmail) {
-    userMenuEmail.textContent = profile?.email || session.user.email || '';
-  }
-
-  if (forceReload) {
-    try {
-      await touchProfileSignIn();
-    } catch (rpcError) {
-      console.warn('Kon laatste inlogmoment niet bijwerken', rpcError);
-    }
-  }
-
-  const shouldReloadData = forceReload || !state.hasLoadedInitialData;
-  if (shouldReloadData) {
-    try {
-      await loadInitialData(profile);
-    } catch (error) {
-      console.error('Kon data niet laden', error);
-      setLoginError('Er ging iets mis bij het laden van de gegevens. Probeer het later opnieuw.');
-      setLoginStatus('');
-      setLoginFormDisabled(false);
-      return;
-    }
-  }
-
-  populateLocationFilters();
-  renderFleet();
-  renderActivity();
-  renderUsers();
-  state.activeTab = 'vloot';
-  applyEnvironmentForRole(profile?.role);
-
-  showAppShell();
-  setLoginStatus('');
-  setLoginError('');
-  setLoginFormDisabled(false);
-  const passwordInput = $('#loginPassword');
-  if (passwordInput) {
-    passwordInput.value = '';
-  }
-}
-
-async function initializeAuth() {
-  try {
-    const { session } = await getCurrentSession();
-    await handleAuthenticatedSession(session, { forceReload: true });
-  } catch (error) {
-    console.error('Initieel ophalen van sessie mislukt', error);
-    showLoginPage();
-  }
-
-  const { data: listener, error: listenerError } = onAuthStateChange(async (event, session) => {
-    const shouldForceReload = event === 'SIGNED_IN';
-    await handleAuthenticatedSession(session, { forceReload: shouldForceReload });
-  });
-
-  if (listenerError) {
-    console.error('Kon Supabase auth listener niet initialiseren', listenerError);
-    return;
-  }
-
-  window.addEventListener('beforeunload', () => {
-    listener?.subscription?.unsubscribe?.();
-  });
-}
-
-async function loadInitialData(profile) {
-  const role = profile?.role ?? null;
-  const membershipsPromise = profile?.id ? fetchFleetMemberships(profile.id) : Promise.resolve([]);
-  const shouldFetchAccountRequests = role === 'Beheerder';
-  const accountRequestsPromise = shouldFetchAccountRequests ? fetchAccountRequests() : Promise.resolve([]);
-  const [locationsResult, fleetResult, usersResult, membershipsResult, accountRequestsResult] = await Promise.allSettled([
-    fetchLocations(),
-    fetchFleet(),
-    fetchUsers(),
-    membershipsPromise,
-    accountRequestsPromise
-  ]);
-
-  let hadError = false;
-
-  if (locationsResult.status === 'fulfilled') {
-    setLocations(locationsResult.value);
-  } else {
-    console.error('Kon locaties niet laden vanuit Supabase.', locationsResult.reason);
-    hadError = true;
-  }
-
-  if (fleetResult.status === 'fulfilled') {
-    setFleet(fleetResult.value);
-  } else {
-    console.error('Kon vloot niet laden vanuit Supabase.', fleetResult.reason);
-    hadError = true;
-  }
-
-  if (usersResult.status === 'fulfilled') {
-    setUsers(usersResult.value);
-  } else {
-    console.error('Kon gebruikers niet laden vanuit Supabase.', usersResult.reason);
-    hadError = true;
-  }
-
-  const unrestricted = role === 'Beheerder' || role === 'Gebruiker';
-  let accessibleFleetIds = unrestricted ? null : new Set();
-
-  if (membershipsResult.status === 'fulfilled') {
-    const memberships = membershipsResult.value || [];
-    if (!unrestricted) {
-      memberships
-        .map(item => item?.fleetId)
-        .filter(Boolean)
-        .forEach(fleetId => accessibleFleetIds.add(fleetId));
-    }
-  } else {
-    console.error('Kon vloottoegang niet laden vanuit Supabase.', membershipsResult.reason);
-    hadError = true;
-  }
-
-  state.accessibleFleetIds = unrestricted ? null : accessibleFleetIds;
-
-  if (accountRequestsResult.status === 'fulfilled') {
-    state.accountRequests = Array.isArray(accountRequestsResult.value) ? accountRequestsResult.value : [];
-  } else {
-    state.accountRequests = [];
-    if (shouldFetchAccountRequests) {
-      console.error('Kon accountaanvragen niet laden vanuit Supabase.', accountRequestsResult.reason);
-      hadError = true;
-    }
-  }
-
-  if (hadError) {
-    showToast('Live data niet beschikbaar – voorbeelddata worden gebruikt.');
-  }
-
-  state.hasLoadedInitialData = true;
-}
+import { wireEvents } from './modules/events.js';
+import { initializeAuth } from './modules/auth.js';
 
 document.addEventListener('DOMContentLoaded', async () => {
   wireEvents();
